==> rmds/1-create_phyloseq.Rmd <==
---
title: "1-Create phyloseq objects"
author: "Siyuan Ma"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    df_print: paged
---
# Overview
- Load .biom files from 16s workflow.
- Load curated metadata. 
- Match samples between otu table and metadata. Create otu cluster/species/genera level phyloseq objects.

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = "../")
```

```{r setup2, include=FALSE}
smar::sourceDir("functions/")
setup("./")
```

# Load .biom files from 16s workflow

```{r load bioms}
print(studies)
l_biom <- list()
for(study in studies) {
  biom <- paste0(dir_processed, 
                 "processed/", 
                 study, 
                 "/16s/all_samples_taxonomy_closed_reference.biom") %>% 
    phyloseq::import_biom()
  if(any(phyloseq::taxa_sums(biom) == 0)) 
    stop("There are zero sum taxa in the data!", study) # all otu clusters should have at least some presence in the data, based on the pipeline
  l_biom[[study]] <- biom
}
```

# Load cuarated metadata, match samples, create phyloseq objects.

```{r create phyloseq}
template <- paste0(dir_processed, "data/template.csv") %>%
  readr::read_csv(col_types = readr::cols(.default = readr::col_character()))
column_setup <- template$var.class %>%
  dplyr::recode("character" = "c",
                "numeric" = "d",
                "integer" = "i") %>%
  paste(collapse = "")
for(study in studies) {
  print(study)
  biom <- l_biom[[study]]
  metadata <- paste0(dir_processed,
                     "processed/",
                     study,
                     "/metadata/metadata.txt") %>%
    readr::read_tsv(col_types = column_setup)

  sample_common <- intersect(metadata$sample_accession_16S,
                             phyloseq::sample_names(biom))
  cat(nrow(metadata),
      phyloseq::nsamples(biom),
      length(sample_common), "\n")

  metadata <- metadata %>%
    dplyr::filter(sample_accession_16S %in% sample_common) %>%
    as.data.frame()
  rownames(metadata) <- metadata$sample_accession_16S

  mat_tax <- smar::tax_table2(biom)
  mat_tax <- cbind(mat_tax, otu = rownames(mat_tax))

  physeq <- phyloseq::phyloseq(
    phyloseq::otu_table(biom)[, sample_common],
    phyloseq::sample_data(metadata[sample_common, ]),
    phyloseq::tax_table(mat_tax))
  print(biom)
  print(physeq)
  dir.create(paste0("data/physeq/", study, "/"),
             recursive = TRUE,
             showWarnings = TRUE)
  physeq_species <- phyloseq::tax_glom(physeq, "Rank7")
  physeq_genera <- phyloseq::tax_glom(physeq_species, "Rank6")
  save(physeq, file = paste0("data/physeq/", study, "/otus.RData"))
  save(physeq_species, file = paste0("data/physeq/", study, "/species.RData"))
  save(physeq_genera, file = paste0("data/physeq/", study, "/genera.RData"))
}
```
