==> rmds/2-QC.Rmd <==
---
title: "2-QC"
author: "Siyuan Ma"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document:
    df_print: paged
---
# Overview
- Filter samples based on read depth
- Filter features based on prevalence and abundance

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = "../")
```

```{r setup2, include=FALSE}
smar::sourceDir("functions/", recursive = FALSE)
setup("./")
dir_output <- "results/2-QC/"
dir.create(dir_output, recursive = TRUE, showWarnings = TRUE)
```

# Load studies

```{r load datasets}
load("data/physeq/otus_prefilter.RData")
load("data/physeq/species_prefilter.RData")
load("data/physeq/genera_prefilter.RData")
```

# Filter samples

```{r filter samples and taxa}
print(c(nsamples(physeq_species_prefilter),
        dplyr::n_distinct(smar::sample_data2(physeq_species_prefilter)$subject_accession),
        nsamples(physeq_genera_prefilter),
        dplyr::n_distinct(smar::sample_data2(physeq_genera_prefilter)$subject_accession)))
physeq_species <- physeq_species_prefilter %>% 
  phyloseq::subset_samples(read_depth > 3000)
physeq_genera <- physeq_genera_prefilter %>% 
  phyloseq::subset_samples(read_depth > 3000)
print(c(nsamples(physeq_species),
        dplyr::n_distinct(smar::sample_data2(physeq_species)$subject_accession),
        nsamples(physeq_genera),
        dplyr::n_distinct(smar::sample_data2(physeq_genera)$subject_accession)))
```

# Filter taxa

```{r examine filtering thresholds}
cutoff_abundance <- c(1e-5, 5e-5, 1e-4, 1e-3)
cutoff_prevalence <- c(0.05, 0.1)
tb_param <-  tidyr::crossing(cutoff_abundance, cutoff_prevalence)
tb_filter <- (1:nrow(tb_param)) %>% 
  purrr::map_dfr(function(i_param) {
    cutoff_abundance <- tb_param$cutoff_abundance[i_param]
    cutoff_prevalence <- tb_param$cutoff_prevalence[i_param]
    list(physeq_genera,
         physeq_species) %>% 
      purrr::map2_dfr(c("genera", "species"), 
                      function(physeq, feature_level) {
                        physeq_rel <- physeq %>% 
                          to_relativeAbundance()
                        tb_presence <- physeq_rel %>% 
                          smar::phyloseq_to_tb() %>% 
                          dplyr::group_by(feature, dataset_name) %>% 
                          dplyr::summarise(
                            n = dplyr::n(),
                            present = sum(abundance > cutoff_abundance) > 
                              n*cutoff_prevalence
                          ) %>% 
                          dplyr::ungroup()
                        tb_presence %>% 
                          dplyr::group_by(feature) %>% 
                          dplyr::summarise(n_study = sum(present)) %>% 
                          dplyr::mutate(cutoff_abundance = cutoff_abundance,
                                        cutoff_prevalence = cutoff_prevalence,
                                        feature_level = feature_level)
                      })
  })
# tb_filter %>%
#   readr::write_tsv(path = paste0(dir_output, "filter.tsv"))
tb_filter_summary <- tb_filter %>%
  dplyr::group_by(feature_level, cutoff_abundance, cutoff_prevalence) %>%
  dplyr::summarise(nfeature = dplyr::n(),
                   nfeature_nonUnique = sum(n_study > 1),
                   nfeature_unique = sum(n_study == 1)) %>%
  dplyr::arrange(feature_level,
                 dplyr::desc(cutoff_abundance), dplyr::desc(cutoff_prevalence)) %>%
  readr::write_tsv(path = paste0(dir_output, "filter_summarise.tsv")) 
```

```{r filter taxa}
physeq_genera <- 
  tb_filter %>% 
  dplyr::filter(feature_level == "genera",
                cutoff_abundance == 5e-5,
                cutoff_prevalence == 0.1,
                n_study >= 1) %>% 
  dplyr::pull(feature) %>% 
  phyloseq::prune_taxa(physeq_genera) %>% 
  smar::prune_taxaSamples()
physeq_species <- 
  tb_filter %>% 
  dplyr::filter(feature_level == "species",
                cutoff_abundance == 5e-5,
                cutoff_prevalence == 0.1,
                n_study >= 1) %>% 
  dplyr::pull(feature) %>% 
  phyloseq::prune_taxa(physeq_species) %>% 
  smar::prune_taxaSamples()
# sanity check
nsamples(physeq_genera)
nsamples(physeq_species)

save(physeq_genera, file = "data/physeq/genera.RData")
save(physeq_species, file = "data/physeq/species.RData")
```
