==> rmds/2.2-examine_metadata.Rmd <==
---
title: "2.2-Examine metadata availability and distribution"
author: "Siyuan Ma"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document:
    df_print: paged
---
# Overview
- Check the availability of metadata across studies
- Visualize distribution

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = "../")
```

```{r setup2, include=FALSE}
smar::sourceDir("functions/", recursive = FALSE)
setup("./")
dir_output <- "results/2.2-examine_metadata/"
dir.create(dir_output, recursive = TRUE, showWarnings = TRUE)
```

# Availability of metadata
```{r check availability}
template <- readr::read_csv(paste0(dir_processed, "data/template.csv"))
load("data/physeq/species_prefilter.RData")
df_metadata <- smar::sample_data2(physeq_species_prefilter)

df_metaLong <- df_metadata %>% 
  dplyr::select(dplyr::one_of(template$col.name)) %>% 
  dplyr::select(-(study_accession:alternative_sample_accession),
                -sample_accession_WGS, -site,
                -(extraction_kit_16S:median_read_length_16S)) %>% 
  tidyr::gather(key = variable, value = value, -dataset_name, -sample_accession_16S) %>% 
  dplyr::left_join(template, by = c("variable" = "col.name")) %>% 
  dplyr::mutate(variable = factor(variable, levels = template$col.name))
df_avail <- df_metaLong %>% 
  dplyr::group_by(dataset_name, variable) %>% 
  dplyr::summarise(is_available = any(!is.na(value)))
p <- df_avail %>% 
  ggplot(aes(x = dataset_name, y = variable, fill = is_available)) + 
  geom_tile() +
  scale_fill_manual(values = c("TRUE" = "black", "FALSE" = "white")) +
  smar::rotate_xaxis(15)
print(p) %>% 
  ggsave(file = paste0(dir_output, "fig_checkAvailabbility.pdf"), .,
       width = 10, height = 10)
```
# Distribution of metadata
```{r distribution of metadata}
df_metaLong <- df_metaLong %>% dplyr::left_join(df_avail, by = c("dataset_name", "variable"))
l_p_disc <- df_metaLong %>% 
  dplyr::filter(var.class == "character", allowedvalues != "*") %>% 
  dplyr::mutate(value_mod = ifelse(!is_available, 
                                   NA,
                                   dplyr::recode(value, .missing = "not available"))) %>% 
  dplyr::group_by(variable) %>% 
  tidyr::nest() %>% 
  dplyr::mutate(p = purrr::map2(variable, data, function(variable, idf) {
    levels_value <- idf$allowedvalues %>% unique() %>% 
      strsplit("|", fixe = TRUE) %>% extract2(1)
    colors_value <- smar::gg_color_hue(levels_value)
    idf %>% 
      dplyr::mutate(value_mod = value_mod %>% 
                      factor(levels = c(levels_value, "not available") %>% 
                               rev)) %>% 
      dplyr::group_by(dataset_name, value_mod) %>% 
      dplyr::summarise(n = sum(!is.na(value_mod))) %>% 
      ggplot(aes(x = dataset_name, y = n, fill = value_mod)) + geom_bar(stat = "identity") +
      scale_fill_manual(values = c(colors_value, "not available" = "darkgray")) +
      ggtitle(variable) + smar::rotate_xaxis(30)
  })) %>% 
  extract2("p")
cowplot::plot_grid(plotlist = l_p_disc, ncol = 2) %>% 
  ggsave(filename = paste0(dir_output, "fig_discDist.pdf"),
         width = 20, height = length(l_p_disc) * 5,
         limitsize = FALSE)
l_p_num <- df_metaLong %>% 
  dplyr::filter(var.class == "numeric") %>% 
  dplyr::group_by(variable) %>% 
  tidyr::nest() %>% 
  dplyr::mutate(p = purrr::map2(variable, data, function(variable, idf) {
    idf %>% 
      dplyr::mutate(value = as.numeric(value)) %>% 
      dplyr::group_by(dataset_name) %>% 
      dplyr::mutate(perc_na = mean(is.na(value)) %>% 
                      multiply_by(100) %>% round(2)) %>%
      dplyr::ungroup() %>% 
      ggplot(aes(x = paste0(dataset_name, "\n", perc_na, "% missing"), 
                 y = value)) + 
      geom_violin() + geom_boxplot(width = 0.1) + 
      ggtitle(variable) + xlab("dataset_name") + smar::rotate_xaxis(30)
  })) %>% 
  extract2("p")
cowplot::plot_grid(plotlist = l_p_num, ncol = 1) %>% 
  ggsave(filename = paste0(dir_output, "fig_numDist.pdf"),
         width = 10, height = length(l_p_num) * 5)
```
