==> rmds/3-adjust_batch.Rmd <==
---
title: "3-Normalize batch and cohort effects"
author: "Siyuan Ma"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    df_print: paged
---
# Overview
- Correct for batch within studies if available.
    - Adjust for disease and sample type if sensible (i.e., not perfectly confounded with batch)
    - This turns out to be just disease for CS-PRISM, LSS-PRISM, MucosalIBD, and RISK
- Correct for study effect overall.
    - Adjust for disease and sample type

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = "../")
```

```{r setup2, include=FALSE}
smar::sourceDir("functions/", recursive = FALSE)
setup("./")
dir_output <- "results/3-adjust_batch/"
dir.create(dir_output, recursive = TRUE, showWarnings = TRUE)
```

# Perform batch-correction on studies where available

```{r adjust batch}
print(studies)
dir_outputChunk <- paste0(dir_output, "batches/")
dir.create(dir_outputChunk, recursive = TRUE, showWarnings = TRUE)
for(feature_level in c("genera", "species")) {
  load(paste0("data/physeq/", feature_level, ".Rdata"))
  physeq <- get(paste0("physeq_", feature_level))
  mat_otu <- smar::otu_table2(physeq)
  metadata <- smar::sample_data2(physeq)

  studies_batch <- metadata %>% 
    dplyr::group_by(dataset_name) %>% 
    dplyr::summarise(has_batch = any(!is.na(batch))) %>% 
    dplyr::filter(has_batch) %>% 
    magrittr::extract2("dataset_name")
  mat_otu_adj <- mat_otu
  l_fit_adonis <- list()
  for(study in studies_batch) {
    samples_tmp <- rownames(metadata)[metadata$dataset_name == study &
                                        !is.na(metadata$batch)] # only adjust for non-missing 
                                                                # batches!
    metadata_tmp <- metadata[samples_tmp, ]
    # adjust for disease and sample type if they are not confounded with batch
    covariates_tmp <- c("disease", "sample_type") %>% 
      purrr::map(function(variable)  {
        if(any(apply(table(metadata_tmp[, variable], metadata_tmp$batch) > 0,
                 2, sum) > 1))
          return(variable)
        return(NULL) 
      }) %>% purrr::reduce(c)
    
    cat("Study", study, "has", length(samples_tmp), "samples with",
        dplyr::n_distinct(metadata[samples_tmp, ]$batch), "batches.\n")
    mat_otu_tmp <- MMUPHin::adjust.batch(feature.abd = mat_otu[, samples_tmp],
                                         batch = "batch",
                                         covariates = covariates_tmp,
                                         data = metadata_tmp,
                                         verbose = FALSE,
                                         diagnostics = TRUE)
    ggsave(filename = paste0(dir_outputChunk, feature_level, "_", study, "_MUPPHin.pdf"),
           width = 10, height = 5)
    
    mat_otu_adj[, samples_tmp] <- mat_otu_tmp
  }
  physeq_adj <- physeq
  otu_table(physeq_adj) <- otu_table(mat_otu_adj, taxa_are_rows = TRUE)
  assign(paste0("physeq_", feature_level, "_adjBatch"), value = physeq_adj)
  save(list = paste0("physeq_", feature_level, "_adjBatch"),
       file = paste0("data/physeq/", feature_level, "_adjBatch.RData"))
}
```

# Perform study adjustment
```{r study adjustment}
for(feature_level in c("genera", "species")) {
  physeq_adj <- get(paste0("physeq_", feature_level, "_adjBatch"))
  mat_otu_adj <- smar::otu_table2(physeq_adj)
  mat_otu_adj <- MMUPHin::adjust.batch(feature.abd = mat_otu_adj,
                                       batch = "dataset_name",
                                       covariates = c("disease", "body_site"),
                                       data = metadata)
  ggsave(filename = paste0(dir_output, feature_level, "_", "MMUPHin.pdf"),
         width = 8,
         height = 4)
  otu_table(physeq_adj) <- otu_table(mat_otu_adj, taxa_are_rows = TRUE)
  assign(paste0("physeq_", feature_level, "_adj"), physeq_adj)
  save(list = paste0("physeq_", feature_level, "_adj"),
       file = paste0("data/physeq/", feature_level, "_adj.RData"))
}
```
