==> rmds/3.1-distance_ordinate.Rmd <==
---
title: "3-Calculate distance measures; ordination visulization"
author: "Siyuan Ma"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    df_print: paged
---
# Overview
These tasks are related and time consuming, so organize together in the same Rmd
- Calcuate distance metrics
- Perform ordination

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = "../")
```

```{r setup2, include=FALSE}
smar::sourceDir("functions/", recursive = FALSE)
setup("./")
dir_output <- "results/3.1-distance_ordinate/"
dir.create(dir_output, recursive = TRUE, showWarnings = TRUE)
```

```{r load phyloseq objects}
load("data/physeq/genera.RData")
load("data/physeq/genera_adjBatch.RData")
load("data/physeq/genera_adj.RData")
load("data/physeq/species.RData")
load("data/physeq/species_adjBatch.RData")
load("data/physeq/species_adj.RData")
```

# These are the distance metrics 

```{r define distances}
distances <- tibble::tibble(distance = c("jaccard", "bray", "jsd"),
                            binary = c(TRUE, FALSE, FALSE))
if(anyDuplicated(distances$distance))
  stop("At this point distance measures should be either binary or continuous!")
```

# Calculate distances

```{r calculate distance}
ncores <- 20
doParallel::registerDoParallel(cores = ncores)
start.time <- Sys.time()
for(i in 1:nrow(distances)) {
  i_distance <- distances$distance[i]
  dir_output_tmp <- paste0("data/distance/", i_distance, "/")
  dir.create(dir_output_tmp, recursive = TRUE, showWarnings = TRUE)
  for(feature_level in c("genera")) {
    # calculate distance on the study-adjusted matrix
    physeq_tmp <- get(paste0("physeq_", feature_level, "_adj")) %>% 
      to_relativeAbundance()
    dist_tmp <- phyloseq::distance(physeq_tmp, 
                                   method = i_distance,
                                   binary = distances$binary[i],
                                   parallel = TRUE)
    assign(paste0("dist_", feature_level, "_adj"), dist_tmp)
    
    # calculate distance on the original matrix
    if(i_distance == "bray") {
      physeq_tmp <- get(paste0("physeq_", feature_level)) %>% 
        to_relativeAbundance()
      dist_tmp <- phyloseq::distance(physeq_tmp, 
                                     method = i_distance,
                                     binary = distances$binary[i],
                                     parallel = TRUE)
      assign(paste0("dist_", feature_level), dist_tmp)
      
      # calculate distance on the batch-adjusted matrix
      physeq_tmp <- get(paste0("physeq_", feature_level, "_adjBatch")) %>% 
        to_relativeAbundance()
      dist_tmp <- phyloseq::distance(physeq_tmp, 
                                     method = i_distance,
                                     binary = distances$binary[i],
                                     parallel=TRUE)
      assign(paste0("dist_", feature_level, "_adjBatch"), dist_tmp)
      
      save(list = paste0("dist_", feature_level),
           file = paste0(dir_output_tmp, "dist_", feature_level, ".RData"))
      save(list = paste0("dist_", feature_level, "_adjBatch"),
           file = paste0(dir_output_tmp, "dist_", feature_level, "_adjBatch.RData"))
    }
    
    save(list = paste0("dist_", feature_level, "_adj"),
         file = paste0(dir_output_tmp, "dist_", feature_level, "_adj.RData"))
  }
}
doParallel::stopImplicitCluster()
print(Sys.time() - start.time)
```

# Ordination

```{r ordinate}
# At this point only ordinate the bray-curtis distance results
distance_to_ordinate <- "bray"
start.time <- Sys.time()
for(i_distance in distance_to_ordinate) {
  dir_output_tmp <- paste0("data/ordinate/", i_distance, "/")
  dir.create(dir_output_tmp, recursive = TRUE, showWarnings = TRUE)
  for(feature_level in c("genera", "species")) {
    load(paste0("data/distance/", 
                i_distance, "/dist_", feature_level,
                ".RData"))
    load(paste0("data/distance/", 
                i_distance, "/dist_", feature_level,
                "_adjBatch.RData"))
    load(paste0("data/distance/", 
                i_distance, "/dist_", feature_level,
                "_adj.RData"))
    # for ordination physeq_genera is okay before metadata is shared
    # ordinate original data
    dist_tmp <- get(paste0("dist_", feature_level))
    ordinate_tmp <- phyloseq::ordinate(physeq_genera, dist_tmp, method = "MDS")
    assign(paste0("ord_", feature_level), ordinate_tmp)
    # ordinate fully adjusted data
    dist_tmp <- get(paste0("dist_", feature_level, "_adj"))
    ordinate_tmp <- phyloseq::ordinate(physeq_genera, dist_tmp, method = "MDS")
    assign(paste0("ord_", feature_level, "_adj"), ordinate_tmp)
    save(list = paste0("ord_", feature_level), 
         file = paste0(dir_output_tmp, "ord_", feature_level, ".RData"))
    save(list = paste0("ord_", feature_level, "_adj"), 
         file = paste0(dir_output_tmp, "ord_", feature_level, "_adj.RData"))
    # for dataset with batch variable, ordinate before and after batch adjustment
    for(study in c("BIDMC-FMT", "CS-PRISM", "LSS-PRISM", "MucosalIBD", "RISK")) {
      physeq_tmp <- phyloseq::subset_samples(physeq_genera, dataset_name %in% study)
      # ordinate original data
      dist_tmp <- get(paste0("dist_", feature_level)) %>% 
        smar::subset_distance(phyloseq::sample_names(physeq_tmp))
      ordinate_tmp <- phyloseq::ordinate(physeq_tmp, dist_tmp, method = "MDS")
      assign(paste0("ord_", feature_level, "_", study), ordinate_tmp)
      dist_tmp <- get(paste0("dist_", feature_level, "_adjBatch")) %>% 
        smar::subset_distance(phyloseq::sample_names(physeq_tmp))
      ordinate_tmp <- phyloseq::ordinate(physeq_tmp, dist_tmp, method = "MDS")
      assign(paste0("ord_", feature_level, "_", study, "_adjBatch"), ordinate_tmp)
      save(list = paste0("ord_", feature_level, "_", study), 
           file = paste0(dir_output_tmp, "ord_", feature_level, "_", study, ".RData"))
      save(list = paste0("ord_", feature_level, "_", study, "_adjBatch"), 
           file = paste0(dir_output_tmp, "ord_", feature_level, "_", study, "_adjBatch.RData"))
    }
  }
}
print(Sys.time() - start.time)
```
