==> rmds/4.1-permanova.Rmd <==
---
title: "4.1-permanova"
author: "Siyuan Ma"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
html_document:
df_print: paged
---
# Overview
Perform PERMANOVA analysis on covariates.

```{r setup, echo = FALSE}                                           
knitr::opts_knit$set(root.dir = normalizePath(".."))
```
```{r setup2, include=FALSE}
smar::sourceDir("functions/", recursive = FALSE)
setup("./")
dir_output <- "results/4.1-permanova/"
dir.create(dir_output, recursive = TRUE, showWarnings = TRUE)
```
# PERMANOVA

```{r define PERMANOVA models}
print(studies)
load("data/physeq/genera.RData")

# create additional variables to test for
metadata <- smar::sample_data2(physeq_genera)
metadata_test <- metadata %>% 
  dplyr::mutate(subtype = disease,
                IBD = disease %>% 
                  dplyr::recode("CD" = "IBD",
                                "UC" = "IBD",
                                "control" = "control"),
                age.cat = 
                  dplyr::case_when(age < 18 ~ "child",
                                   !is.na(age) ~ "adult",
                                   TRUE ~ NA_character_))
rownames(metadata_test) <- rownames(metadata)

study_longitudinal <- tb_1$`Longitudinal sampling?` == "Yes"
names(study_longitudinal) <- tb_1$study_full_name

variable_IBD.cat <- c("L.cat", "B.cat", "E.cat")
variable_treatment <- c("antibiotics", "immunosuppressants", "steroids", "mesalamine_5ASA")
variable_demo <- c("age.cat", "age_at_diagnosis.cat", "gender", "race")

# one study * variable per-row tb
tb_permanova <- tibble::tibble(
  variable = c("dataset_name", "dataset_name", "batch", "batch",
               "sample_type", "body_site",
               "IBD", "disease",
               "L.cat", "B.cat", "E.cat",
               "antibiotics", "immunosuppressants", "steroids", "mesalamine_5ASA",
               "age.cat", "age_at_diagnosis.cat", "gender", "race",
               "subject_accession"),
  physeq = c("original", "adjusted", "original", "batch_adjusted",
            "adjusted", "adjusted",
            "adjusted", "adjusted",
            "adjusted", "adjusted", "adjusted",
            "adjusted", "adjusted", "adjusted", "adjusted",
            "adjusted", "adjusted", "adjusted", "adjusted",
            "adjusted")) %>%
  tidyr::crossing(study = c("all", studies)) %>%
  dplyr::left_join(tb_1, by = c("study" = "study_full_name")) %>%
  dplyr::mutate(study_longitudinal = `Longitudinal sampling?` == "Yes") %>%
  dplyr::select(study, study_longitudinal, variable, physeq)

# specifying models for each study-variable pair
tb_permanova <- tb_permanova %>%
  # specify variable type (per-sample or per-subject), 
  # covariates to adjust for (if any), 
  # and covariate type (per-sample or per-subject)
  dplyr::mutate(
    variable_type = dplyr::case_when( # variable is per sample or per subject?
      variable %in% c("dataset_name", "IBD", "disease", 
                      variable_IBD.cat, variable_demo) ~ 
        "subject",
      variable %in% c("batch", "sample_type", "body_site",
                      variable_treatment, "subject_accession") ~ # subject needs to be tested 
        # per-sample!
        "sample"),
    variable_adjust = dplyr::case_when( # does model adjust for additional covariates?
      variable == "body_site" ~ "sample_type",
      variable == "disease" ~ "IBD",
      variable %in% c(variable_IBD.cat, variable_treatment, 
                      "age_at_diagnosis.cat") ~ "disease",
      TRUE ~ NA_character_
    ),
    variable_adjust_type = dplyr::case_when( # varialbe adjust is per sample or per subject?
      variable_adjust %in% c("sample_type") ~ "sample",
      variable_adjust %in% c("IBD", "disease") ~ "subject",
      TRUE ~ NA_character_
    )
  ) %>%
  # specify, depending on study-variable pair, which version of PERMANOVA wrapper,
  # i.e., permutation scheme to run (study, subject, or sample version)
  dplyr::mutate(permanova_call = dplyr::case_when(
    # Testing all studies, usually run the study version
    study == "all" & !(variable %in% c("dataset_name", "subject_accession")) ~ 
      "study",
    # Testing all studies but for study and subject effect. Special case and 
    # see inside for loop for implementation
    study == "all" & (variable %in% c("dataset_name", "subject_accession")) ~ 
      "subject",
    # Testing longitudinal study, usually run subject version
    study != "all" & study_longitudinal & !(variable %in% c("batch", "subject_accession")) ~ 
      "subject",
    # Testing longitudinal study, but for batch and subject effects, run
    # per-sample (i.e., free) permutation
    study != "all" & study_longitudinal & (variable %in% c("batch", "subject_accession")) ~ 
      "sample",
    # Testing cross-sectional study
    study != "all" & !study_longitudinal ~ 
      "sample",
    TRUE ~ NA_character_
  )) %>% 
  dplyr::group_by(study, variable) %>%
  # subsetting metadata to the specific study for each study-variable pair
  dplyr::mutate(metadata = dplyr::case_when(
    study == "all" ~ list(metadata_test),
    TRUE ~ list(subset(metadata_test, dataset_name %in% study))
  )) %>% 
  # Within inidivudal study, create artificial "missingness" indicator covariate, 
  # which is adjusted for as covariate in the PERMANOVA model
  dplyr::mutate(
    # percentage of missingness for the variable
    variable_missing = mean(is.na(metadata[[1]][, variable])),
    # only fill in variable if missingness is greater than zero and less than one
    variable_na_fill = variable_missing < 1 & variable_missing > 0,
    # the number of non-missing categories for the variable
    variable_ncat = dplyr::n_distinct(metadata[[1]][, variable[1]], na.rm = TRUE),
    # for not meaningful cases not run PERMANOVA
    variable_fit = dplyr::case_when( 
      variable_missing == 1 ~ FALSE,
      variable_ncat <= 1 ~ FALSE,
      study == "all" & variable == "batch" ~ FALSE,
      study != "all" & variable == "dataset_name" ~ FALSE,
      study_longitudinal %in% FALSE & variable == "subject_accession" ~ FALSE,
      TRUE ~ TRUE),
    # number of categories for the adjustment variable, if present
    variable_adjust_ncat = 
      ifelse(is.na(variable_adjust), 
             NA_integer_,
             dplyr::n_distinct(metadata[[1]][, variable_adjust[[1]]])),
    # if no covariate, or number of categories is only 1, then don't fit covariate
    variable_adjust_fit = dplyr::case_when(
      !variable_fit ~ FALSE,
      is.na(variable_adjust) ~ FALSE,
      variable_adjust_ncat <= 1 ~ FALSE,
      TRUE ~ TRUE),
    # not always fit missigness as covariate
    variable_na_fit = dplyr::case_when(
      # obviously, if variabe itself is not fitted
      !variable_fit ~ FALSE,
      # if no filling in is required
      !variable_na_fill ~ FALSE,
      # if covariate is present, and perfectly encompass missingness
      # (such as L.cat is and only is missing for CD, and disease is adjusted for)
      variable_adjust_fit & 
        ifelse(variable_adjust_fit,
               all(apply(table(metadata[[1]][, variable_adjust],
                               is.na(metadata[[1]][, variable])) > 0,
                         1,
                         sum) <= 1),
               FALSE) ~ 
        FALSE,
      TRUE ~ TRUE)
  ) %>% 
  dplyr::select(study, study_longitudinal,
                physeq,
                variable, variable_type, variable_fit,
                variable_adjust, variable_adjust_type, variable_adjust_fit,
                variable_na_fill, variable_na_fit,
                permanova_call) %>% 
  dplyr::ungroup() %>% 
  dplyr::arrange(study, variable)
save(tb_permanova, file = "results/4.1-permanova/permanova_config.RData")
```

```{r run PERMANOVA}
n_permutations <- 1999 # set # permuations for PERMANOVA
ncores <- 20
load("data/distance/bray/dist_genera.RData")
load("data/distance/bray/dist_genera_adj.RData")
load("data/distance/bray/dist_genera_adjBatch.RData")
start.time <- Sys.time()
for(i_run in (1:nrow(tb_permanova))) {
  i_tb_permanova <- tb_permanova[i_run, ]
  
  if(!i_tb_permanova$variable_fit) {
    adonis_fit <- NULL
    save(adonis_fit, file = paste0("results/4.1-permanova/adonis_fit_",
                                   i_run, ".RData"))
    next
  }
  
  if(i_tb_permanova$study == "all") {
    i_metadata_test <- metadata_test
  } else {
    i_metadata_test <- subset(metadata_test, dataset_name %in% i_tb_permanova$study)
  }
  
  # specify study and subject variables for block permutation
  i_study <- i_metadata_test$dataset_name
  names(i_study) <- rownames(i_metadata_test)
  if(i_tb_permanova$study == "all" & i_tb_permanova$variable == "subject") {
    i_subject <- i_metadata_test$dataset_name # for subject R2 in all studies,
    # permute "subject" freely but within studies
  } else {
    i_subject <- i_metadata_test$subject_accession 
    # note, this includes the case when test is for study effects across studies, 
    # where "study" is treated as a per-subject variable and permuted along with subject
    # block structure
  }
  names(i_subject) <- rownames(i_metadata_test)
  
  # fill in NA values (because adonis does not allow for missing variables) if needed
  # and add missingness indicator
  if(i_tb_permanova$variable_na_fill)  {
    i_metadata_test$na <- is.na(i_metadata_test[, i_tb_permanova$variable])
    i_metadata_test[, i_tb_permanova$variable] <- 
      fill_na(i_metadata_test[, i_tb_permanova$variable])
  }
  
  # variables to include in the PERMANOVA model, including the variable to test for 
  # itself, covariate to adjust for if present, and missingness indicator if required
  variables_fit <- i_tb_permanova$variable
  if(i_tb_permanova$variable_adjust_fit)
    variables_fit <- c(i_tb_permanova$variable_adjust, variables_fit)
  if(i_tb_permanova$variable_na_fit)
    variables_fit <- c("na", variables_fit)
  # is any of the variable, covariate, and missingness indicator per-subject?
  variables_fit_subject <- c(
    i_tb_permanova$variable_adjust[
      i_tb_permanova$variable_adjust_fit &
        i_tb_permanova$variable_adjust_type %in% "subject"], 
    "na"[
      i_tb_permanova$variable_na_fit &
        i_tb_permanova$variable_type == "subject"], 
    i_tb_permanova$variable[i_tb_permanova$variable_type == "subject"]
  )
  # the rest are per-sample variables
  variables_fit_sample <- setdiff(variables_fit, variables_fit_subject)
  # create per-subject (one subject a row) data frame if any of the variables are per-subject
  if(length(variables_fit_subject > 0)) {
    i_subject_data <- i_metadata_test[!duplicated(i_metadata_test$subject_accession), 
                                      variables_fit_subject,
                                      drop = FALSE]
    rownames(i_subject_data) <- 
      i_metadata_test$subject_accession[!duplicated(i_metadata_test$subject_accession)]
  } else {
    i_subject_data <- NULL
  }
  # create per-sample data frame if any of the variables are per-sample
  if(length(variables_fit_sample > 0)) {
    i_sample_data <- i_metadata_test[, variables_fit_sample, drop = FALSE]
  } else {
    i_sample_data <- NULL
  }
  
  # speicfy distance objects to test 
  if(i_tb_permanova$physeq == "original") 
    i_dist <- smar::subset_distance(dist_genera, rownames(i_metadata_test))
  if(i_tb_permanova$physeq == "adjusted") 
    i_dist <- smar::subset_distance(dist_genera_adj, rownames(i_metadata_test))
  if(i_tb_permanova$physeq == "batch_adjusted") {
    i_dist <- smar::subset_distance(dist_genera_adjBatch, rownames(i_metadata_test))
  }
  
  # run the models
  if(i_tb_permanova$permanova_call == "study")
    adonis_fit <- 
    PERMANOVA_repeat_measures_meta(D = i_dist,
                                   study = i_study,
                                   study_longitudinal = study_longitudinal,
                                   subject = i_subject,
                                   subject_data = i_subject_data,
                                   sample_data = i_sample_data,
                                   metadata_order = variables_fit,
                                   permutations = n_permutations,
                                   ncores = ncores)
  if(i_tb_permanova$permanova_call == "subject")
    adonis_fit <- 
    PERMANOVA_repeat_measures(D = i_dist,
                              subject = i_subject,
                              subject_data = i_subject_data,
                              sample_data = i_sample_data,
                              metadata_order = variables_fit,
                              permutations = n_permutations,
                              ncores = ncores)
  if(i_tb_permanova$permanova_call == "sample")
    adonis_fit <- 
    vegan::adonis(i_dist ~ ., 
                  data = i_metadata_test[, variables_fit, drop = FALSE],
                  permutations = n_permutations,
                  parallel = ncores)
  save(adonis_fit, file = paste0("results/4.1-permanova/adonis_fit_",
                                 i_run, ".RData"))
}
print(Sys.time() - start.time)
```
