==> rmds/4.4-interaction.Rmd <==
---
title: "4.4-interaction"
author: "Siyuan Ma"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
html_document:
df_print: paged
---
# Overview
Interaction modelling wrt sample type and disease subtype

```{r setup, echo = FALSE}                                           
knitr::opts_knit$set(root.dir = normalizePath(".."))
```

```{r setup2, include=FALSE}
smar::sourceDir("functions/", recursive = FALSE)
setup("./")
dir_output <- "results/4.4-interaction/"
dir.create(dir_output, recursive = TRUE, showWarnings = TRUE)
```

```{r load data}
source("assets/ma_tests.R")
load("data/physeq/species_adj.RData")
metadata <- smar::sample_data2(physeq_species_adj)
```

# Effect modification of sample type/disease status?
```{r Maaslin2 effect modification}
metadata_test <- metadata %>% 
  dplyr::mutate(race_fill = fill_na(race),
                gender_fill = fill_na(gender),
                age.cat_fill = 
                  dplyr::case_when(age < 18 ~ "child",
                                   !is.na(age) ~ "adult",
                                   TRUE ~ NA_character_) %>% fill_na,
                study_site = paste(dataset_name, 
                                   sample_type, 
                                   sep = "_"),
                study_site_disease = paste(dataset_name, 
                                           sample_type, 
                                           disease, sep = "_"),
                subject_accession_longitudinal =
                  ifelse(dataset_name %in%
                           studies_longitudinal,
                         subject_accession,
                         NA_character_)) 

df_moderator_site <- metadata_test %>% 
  dplyr::select(study_site, sample_type, dataset_name) %>% 
  dplyr::group_by(study_site) %>% 
  dplyr::summarise_all(function(x) unique(x)) %>% 
  dplyr::mutate(sample_type = factor(sample_type)) %>% 
  as.data.frame() %>% tibble::column_to_rownames("study_site") 
df_moderator_siteDisease <- metadata_test %>% 
  dplyr::select(study_site_disease, sample_type, disease, dataset_name) %>% 
  dplyr::filter(disease %in% c("CD", "UC")) %>% 
  dplyr::group_by(study_site_disease) %>% 
  dplyr::summarise_all(function(x) unique(x)) %>% 
  dplyr::mutate(sample_type = factor(sample_type),
                disease = factor(disease)) %>% 
  as.data.frame() %>% tibble::column_to_rownames("study_site_disease")

for(feature_level in "genera") {
  dir.create(paste0(dir_output, feature_level, "/"))
  l_result_disease <- list()
  for(test in tests_disease) {
    print(test)
    test_param <- l_tests[[test]]
    
    load(paste0("results/4.2-maaslin2/", feature_level, "/",
                test, "/result.RData"))
    result_moderator <- fit_rma.mod(l.Maaslin.fit = result$result$ind.results, 
                                    data.moderator = df_moderator_site,
                                    method = "REML",
                                    rma.threshold = 1e-6,
                                    rma.maxiter = 2000) %>% 
      dplyr::mutate(test = test)
    readr::write_tsv(result_moderator, 
                     path = paste0(dir_output, feature_level, "/",
                                   test, "_moderator_results.tsv"))
    plot_metaAnalysis_moderator(fit.lm.meta = result$result,
                                fit.rma.mod = result_moderator,
                                data.moderator = df_moderator_site[, -2, drop = FALSE],
                                test = test,
                                q.cutoff = 0.25,
                                directory = paste0(dir_output, feature_level, "/"))
  }
  for(test in tests_treatment) {
    print(test)
    test_param <- l_tests[[test]]
    
    load(paste0("results/4.2-maaslin2/", feature_level, "/",
                test, "/result.RData"))
    result_moderator <- fit_rma.mod(l.Maaslin.fit = result$result$ind.results, 
                                    data.moderator = df_moderator_siteDisease,
                                    method = "REML",
                                    rma.threshold = 1e-6,
                                    rma.maxiter = 2000) %>% 
      dplyr::mutate(test = test)
    readr::write_tsv(result_moderator, 
                     path = paste0(dir_output, feature_level, "/",
                                   test, "_moderator_results.tsv"))
    plot_metaAnalysis_moderator(fit.lm.meta = result$result,
                                fit.rma.mod = result_moderator,
                                data.moderator = df_moderator_siteDisease[, -3, drop = FALSE],
                                test = test,
                                q.cutoff = 0.25,
                                directory = paste0(dir_output, feature_level, "/"))
  }
}
```
