==> rmds/6-unsupervised_continuous.Rmd <==
---
title: "6-unsupervised_continuous"
author: "Siyuan Ma"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
html_document:
df_print: paged
---
# Overview
Simulation study of unsupervised continuous structure discovery

```{r setup, echo = FALSE}                                           
knitr::opts_knit$set(root.dir = normalizePath(".."))
```
```{r setup2, message=FALSE, warning=FALSE, echo = FALSE}
smar::sourceDir("functions/", recursive = FALSE)
setup("./")
dir_output <- "results/6-unsupervised_continuous/"
dir.create(dir_output, recursive = TRUE, showWarnings = FALSE)
```

# load and prepare data
```{r load and prepare data}
load(paste0("data/physeq/genera_adj.RData"))
metadata <- smar::sample_data2(physeq_genera_adj)

metadata <- metadata %>% 
  dplyr::mutate(study_site = paste(dataset_name, 
                                   sample_type, 
                                   sep = "_"),
                study_disease = paste(dataset_name, 
                                      disease, 
                                      sep = "_"),
                study_site_disease = paste(dataset_name, 
                                           sample_type, 
                                           disease, sep = "_"),
                is_baseline_withCS = 
                  dplyr::case_when(is.na(is_baseline) ~ TRUE,
                                   is_baseline ~ TRUE,
                                   !is_baseline ~ FALSE))
metadata %>% dplyr::filter(disease %in% c("CD", "UC"), is_baseline_withCS) %>% 
  dplyr::pull(study_site_disease) %>% table()
study_site_disease_training <- c(
  "CS-PRISM_stool_CD",
  "CS-PRISM_stool_UC",
  "Pouchitis_biopsy_UC",
  "PROTECT_biopsy_UC",
  "PROTECT_stool_UC",
  "RISK_biopsy_CD",
  "RISK_stool_CD"
)
metadata <- metadata %>% 
  dplyr::mutate(study_site_disease =
                  ifelse(dataset_name == "PROTECT" & !is_baseline_withCS,
                         paste0(study_site_disease, "(not baseline)"),
                         study_site_disease))
rownames(metadata) <- metadata$sample_accession_16S
```

```{r training}
for(feature_level in c("genera")) {
  physeq <- get(paste0("physeq_", feature_level, "_adj")) %>% 
    to_relativeAbundance()
  
  metadata <- metadata[sample_names(physeq), ]
  sample_data(physeq) <- metadata
  dir_outputChunk <- paste0(dir_output, feature_level, "/")
  dir.create(dir_outputChunk, recursive = TRUE, showWarnings = TRUE)
  
  physeq_training <- physeq %>% 
    phyloseq::subset_samples(study_site_disease %in% study_site_disease_training)
  for(cor.cutoff in seq(0.5, 0.9, by = 0.01)) {
    dir_outputTmp <- paste0(dir_outputChunk, "cor_cutoff_", cor.cutoff, "/")
    dir.create(dir_outputTmp, recursive = TRUE, showWarnings = TRUE)
    pdf(paste0(dir_outputTmp, "network_communities.pdf"), width = 8, height = 8)
    fit_continuous <- MMUPHin::continuous.discover(feature.abd = smar::otu_table2(physeq_training),
                                                   batch = "study_site_disease",
                                                   data = smar::sample_data2(physeq_training),
                                                   normalization = "NONE",
                                                   transform = "AST",
                                                   var.perc.cutoff = 0.8,
                                                   cor.cutoff = cor.cutoff,
                                                   plot.clustered.network = TRUE,
                                                   diagnostics = FALSE,
                                                   verbose = TRUE)
    dev.off()
    readr::write_tsv(as.data.frame(fit_continuous$consensus.loading) %>% 
                       tibble::rownames_to_column("feature"),
                     path = paste0(dir_outputTmp, "avg_loading.tsv"))
    save(fit_continuous, file = paste0(dir_outputTmp, "fit_continuous.RData"))
  }
}
```

```{r visualize and validate loading}
metadata <- metadata %>% 
  dplyr::mutate(batch_validation =
                  dplyr::case_when(
                    control == "nonIBD" ~ "nonIBD",
                    control == "HC" ~ "HC",
                    TRUE ~ dataset_name))
rownames(metadata) <- metadata$sample_accession_16S
for(feature_level in c("genera")) {
  dir_outputChunk <- paste0(dir_output, feature_level, "/")
  physeq <- get(paste0("physeq_", feature_level, "_adj")) %>% 
    to_relativeAbundance()
  metadata <- metadata[sample_names(physeq), ]
  sample_data(physeq) <- metadata
  
  batch <- metadata$batch_validation
  lvl.batch <- unique(batch)
  feature.pca <- asin(sqrt(smar::otu_table2(physeq)))
  pca.all <- lapply(lvl.batch, function(lvl)
  {
    pc <- feature.pca[, batch == lvl]
    dat.pca <- prcomp(t(pc))
    return(dat.pca)
  })
  
  data.loadings <- lapply(pca.all, function(x)
  {
    loadings <- x$rotation[,1:7]
    return(loadings)
  })
  
  tb_sampleSize <- metadata %>%
    dplyr::group_by(batch_validation) %>%
    dplyr::summarise(n = n())

  for(cor.cutoff in seq(0.5, 0.9, by = 0.01)) {
    tb_loading <- paste0(dir_outputChunk, "cor_cutoff_", 
                         cor.cutoff, "/avg_loading.tsv") %>% 
      readr::read_tsv() %>% as.data.frame() %>% 
      tibble::column_to_rownames("feature")
    n_axes <- min(5, ncol(tb_loading))
    tb_cor <- data.loadings %>% 
      purrr::map2_dfr(lvl.batch, function(data_loading, lvl) {
        abs(t(data_loading) %*% as.matrix(tb_loading[, 1:n_axes, drop = FALSE])) %>% 
          tibble::as_tibble(rownames = "PC") %>% 
          dplyr::mutate(batch = lvl)
      }) %>% 
      readr::write_tsv(paste0(dir_outputChunk, "cor_cutoff_", 
                              cor.cutoff, "/validation.tsv"))
    p_validation <- tb_cor %>% 
      tidyr::gather(key = Cluster, value = correlation, -PC, -batch) %>% 
      dplyr::group_by(Cluster, batch) %>% 
      dplyr::arrange(desc(correlation)) %>% 
      dplyr::slice(1) %>% dplyr::ungroup() %>% 
      dplyr::left_join(tb_sampleSize, by = c("batch" = "batch_validation")) %>% 
      dplyr::mutate(batch_n = paste0(batch, " (", n, ")"),
                    batch_class = dplyr::case_when(
                      batch %in% c("PROTECT_biopsy", 
                                   "PROTECT_stool", 
                                   "RISK_biopsy", 
                                   "RISK_stool",
                                   "CS-PRISM_biopsy",
                                   "CS-PRISM_stool",
                                   "Pouchitis_biopsy") ~ "Training",
                      batch == "nonIBD" ~ "nonIBD",
                      batch == "HC" ~ "HC",
                      TRUE ~ "Validation"
                    )) %>% 
      ggplot(aes(x = batch_n, y = correlation, 
                 color = batch_class,
                 label = PC)) +
      geom_point() +
      ggrepel::geom_text_repel() +
      facet_grid(Cluster~.) +
      geom_hline(yintercept = cor.cutoff, color = "red", linetype = "dashed") +
      scale_color_manual(values = c("Training" = "red", 
                                    "Validation" = "black",
                                    "nonIBD" = "darkgrey",
                                    "HC" = "grey"), guide = FALSE) +
      coord_flip()
    ggsave(filename = paste0(dir_outputChunk, 
                             "cor_cutoff_", cor.cutoff, 
                             "/validation.pdf"),
           p_validation, 
           width = 8,
           height = n_axes*5)
  }
}
```

```{r validation 2}
# specifically validate loadings for 0.65 cutoff
pc.max <- 3
R <- 200
physeq <- get(paste0("physeq_genera_adj")) %>% 
  to_relativeAbundance()
metadata <- metadata[sample_names(physeq), ]
sample_data(physeq) <- metadata
batch <- metadata$batch_validation
lvl.batch <- unique(batch)
feature.pca <- asin(sqrt(smar::otu_table2(physeq)))

tb_loading <- readr::read_tsv("results/6-unsupervised_continuous/genera/cor_cutoff_0.65/avg_loading.tsv") %>% 
  as.data.frame() %>% 
  tibble::column_to_rownames("feature") %>% 
  extract(1:2)

set.seed(1)
l_tb.validation <- lvl.batch %>% 
  purrr::map(function(i.batch) {
    dat <- feature.pca[, batch == i.batch]
    pca.original <- prcomp(t(dat))
    pca.bootstrap <- (1:R) %>% 
      purrr::map(function(r) dat[, sample.int(n = ncol(dat),replace = TRUE)] %>% 
                   t() %>% prcomp())
    i.tb.validation <- c(list(pca.original), pca.bootstrap) %>% 
      purrr::map2_dfr(c("Original", rep("Bootstrap", R)),
                      function(fit.pca, class_data) {
                        t(fit.pca$rotation[, 1:pc.max]) %*% as.matrix(tb_loading) %>% 
                          abs() %>% 
                          apply(2, max) %>% 
                          tibble::tibble(InnerProd = .,
                                         Cluster = c("Cluster 1", "Cluster 2"),
                                         class_data = class_data,
                                         batch = i.batch)
                      })
    return(i.tb.validation)
  })

# randomized validation
mat_abd <- smar::otu_table2(physeq_genera_adj)
n_randomize <- 473 # median sample size
pca.randomization <- (1:R) %>% 
  purrr::map(function(r) 
    sample.int(n = ncol(mat_abd), size = n_randomize) %>% 
      sapply(function(i) mat_abd[sample.int(n = nrow(mat_abd)), i]) %>% 
      apply(2, MMUPHin:::tss) %>% 
      sqrt() %>% asin() %>% t() %>% 
      prcomp())
tb.validation.random <- pca.randomization %>% 
  purrr::map_dfr(function(fit.pca) {
                      t(fit.pca$rotation[, 1:pc.max]) %*% as.matrix(tb_loading) %>% 
                        abs() %>% 
                        apply(2, max) %>% 
                        tibble::tibble(InnerProd = .,
                                       Cluster = c("Cluster 1", "Cluster 2"),
                                       class_data = "Bootstrap",
                                       batch = "Neg. control")
                    })
tb.validation.random <- tb.validation.random %>% 
  dplyr::group_by(Cluster, batch) %>% 
  dplyr::summarise(InnerProd = mean(InnerProd)) %>% 
  dplyr::ungroup() %>% 
  dplyr::mutate(class_data = "Original") %>% 
  dplyr::bind_rows(tb.validation.random)

tb.validation.results <- c(l_tb.validation, list(tb.validation.random)) %>% 
  purrr::reduce(rbind) %>% 
  readr::write_tsv(paste0(dir_outputChunk, "validation.tsv"))
```

```{r visualize scores}
tb_loading <- readr::read_tsv("results/6-unsupervised_continuous/genera/cor_cutoff_0.65/avg_loading.tsv")

mat_otu_transformed <- physeq_genera_adj %>% 
  to_relativeAbundance() %>% 
  smar::otu_table2() %>% 
  sqrt() %>% asin() %>% t()
# sanity check
# all(rownames(mat_otu_transformed) == rownames(metadata))
mat_scores <- studies %>% 
  purrr::map(function(study) {
    mat_otu_transformed[metadata$dataset_name == study, ] %>% 
      apply(2, function(x) x - mean(x)) %>% 
      multiply_by_matrix(as.matrix(tb_loading[, c("Cluster_1", "Cluster_2")])) %>% 
      as.data.frame()
  }) %>% purrr::reduce(rbind)
mat_scores <- mat_scores[phyloseq::sample_names(physeq_genera_adj), ]
readr::write_tsv(mat_scores %>% 
                   set_colnames(c("Score (dysbiosis)", "Score (phyla tradeoff)")) %>% 
                   tibble::rownames_to_column("sample"), 
                 paste0(dir_outputChunk, "scores.tsv"))

tb_loading <- smar::tax_table2(physeq_genera_adj) %>% 
  as.data.frame %>% 
  tibble::rownames_to_column("feature") %>% 
  dplyr::right_join(tb_loading, by = "feature") %>%
  dplyr::mutate()
p1 <- tb_loading %>% 
  dplyr::mutate(feature_plot = betterGeneraNames(Rank4, Rank5, Rank6)) %>% 
  dplyr::arrange(desc(abs(Cluster_1))) %>% 
  dplyr::slice(1:20) %>% 
  dplyr::arrange(Cluster_1) %>% 
  dplyr::mutate(feature_plot = factor(feature_plot, levels = unique(feature_plot))) %>% 
  ggplot(aes(x = feature_plot, y = Cluster_1)) +
  geom_bar(stat = "identity") +
  coord_flip()
ggsave(filename = paste0(dir_outputChunk, "loadings1.pdf"), p1,
       width = 6, height = 4)
p2 <-tb_loading %>% 
  plyr::mutate(feature_plot = betterGeneraNames(Rank4, Rank5, Rank6)) %>% 
  dplyr::arrange(desc(abs(Cluster_2))) %>% 
  dplyr::slice(1:20) %>% 
  dplyr::arrange(Cluster_2) %>% 
  dplyr::mutate(feature_plot = factor(feature_plot, levels = unique(feature_plot))) %>% 
  dplyr::mutate(Phylum = Rank2) %>% 
  ggplot(aes(x = feature_plot, y = Cluster_2)) +
  geom_bar(stat = "identity", aes(color = Phylum), fill = "white") +
  coord_flip()
ggsave(filename = paste0(dir_outputChunk, "loadings2.pdf"), p2,
       width = 8, height = 4)

load("data/ordinate/bray/ord_genera_adj.RData")

metadata_scores <- smar::sample_data2(physeq_genera_adj) %>% 
  cbind(mat_scores) %>% 
  cbind(ord_genera_adj$vectors[, 1:4])

p_density <- metadata_scores %>% 
  dplyr::mutate(Disease = dplyr::case_when(
    disease == "CD" ~ "CD",
    disease == "UC" ~ "UC",
    control == "nonIBD" ~ "Non-IBD control",
    TRUE ~ "Healthy control"
  ) %>% factor(levels = c("CD", "UC","Non-IBD control", "Healthy control"))) %>% 
  ggplot(aes(x = Cluster_1)) +
  scale_fill_manual(values = c(smar::gg_color_hue(c("CD", "UC","Non-IBD control")), 
                               c("Healthy control"="grey"))) +
  scale_color_manual(values = c(smar::gg_color_hue(c("CD", "UC","Non-IBD control")), 
                                c("Healthy control"="grey"))) +
  geom_density(aes(fill = Disease, color = Disease), alpha = 0.2)
ggsave("results/6-unsupervised_continuous/genera/density.pdf",
       p_density,
       width = 6, height = 4)

p_score1 <- metadata_scores %>% 
  dplyr::mutate(Disease = dplyr::case_when(
    disease == "CD" ~ "CD",
    disease == "UC" ~ "UC",
    control == "nonIBD" ~ "Non-IBD control",
    TRUE ~ "Healthy control"
  ) %>% factor(levels = c("CD", "UC","Non-IBD control", "Healthy control"))) %>% 
  ggplot(aes(x = Axis.1, y = Axis.2, color = Cluster_1)) +
  geom_point() +
  facet_grid(.~Disease) +
  scale_color_gradient2(low = "blue", high = "red", mid = "white", midpoint = 0)
ggsave("results/6-unsupervised_continuous/genera/Score1.pdf", p_score1, width = 13, height = 3)
p_score2 <- metadata_scores %>% 
  dplyr::mutate(Disease = dplyr::case_when(
    disease == "CD" ~ "CD",
    disease == "UC" ~ "UC",
    control == "nonIBD" ~ "Non-IBD control",
    TRUE ~ "Healthy control"
  ) %>% factor(levels = c("CD", "UC","Non-IBD control", "Healthy control"))) %>% 
  ggplot(aes(x = Axis.1, y = Axis.2, color = Cluster_2)) +
  geom_point() +
  facet_grid(.~Disease) +
  scale_color_gradient2(low = "blue", high = "red", mid = "white", midpoint = 0)
ggsave("results/6-unsupervised_continuous/genera/Score2.pdf", p_score2, width = 13, height = 3)
```
